- Change battery message decoding - Change area_id, executor_id and loc_id from uint8 to uint32 - Fix calling planner error due to no plan returned

+++ urs_wearable/include/urs_wearable/knowledge_base.h
@@ -20,8 +20,8 @@ public:
   // We use uint8_t here to match with the type of 'type' in Predicate.msg
   typedef std::uint8_t predicate_type;
 
-  // We use uint8_t here to match with the type of 'executor_id' in Feedback.msg
-  typedef std::uint8_t executor_id_type;
+  // We use uint8_t here to match with the type of 'executor_id' in Feedback.msg, SetGoal.srv
+  typedef std::uint32_t executor_id_type;
 
   LocationTable loc_table_;
